requirements.txt
streamlit
python-dotenv
requests
beautifulsoup4
openai
pandas
